Remove early exit in licensecheck
# Motivation

When onboarding a new repo it is often the case that new file extensions need to be added to this repo. Currently, we fail early when we encounter an unknown extension.

# Modification

This PR changes to produce an error when we encounter an unknown extension but continue.

# Result

Makes it easier to onboard new repos.