Add optional unacceptable language ignore file
Some of our repos contain code that is pulled in from third parties such as swift-crypto that vendors boringssl. This code can contain unacceptable language. We cannot disable this currently since the only way to do this is by adding comments.

To support such use-cases this PR extends the unacceptable-language script to also look for an optional `.unacceptablelanguageignore` file at the root of the repo.