Add ability to include swift versions
In some cases, such as swiftly, there's only one particular version
of swift that is being targeted at a time. It's problematic trying
to exclude all of the others in the default list, and have to adjust
when the workflow gets changed, making sure to exclude any additions.
Add the ability for a repo like swiftly to set the list to explicit
version(s) instead of excluding everything except that one version.