Allow using GITHUB_TOKEN downstream
Looking at https://docs.github.com/en/actions/writing-workflows/choosing-what-your-workflow-does/accessing-contextual-information-about-workflow-runs#github-context
the `github.token` is only available withins the `steps`. Of the common
workflow, but the jobs extending it do not have access. The vscode-swift
extension publishes a build of the extension in a separate job, and we
want to be able to download it within the common workflow so we can test
what is actually might be released.