Add GitHub action that automatically creates a PR to merge main into a release branch
The typical use case for this is in the first period after Swift has cut release branches. Some repositories want to include most changes from `main` also in the release branch. When this job is set up, it can automatically create PRs to merge `main` into the release branch. Maintainers of the package can then inspect the changes to ensure that they are not too risky for the release branch. We will also run the normal PR testing on these changes, ensuring that these modifications don't break the build.