Add workflow to test the performance of a package
This adds the infrastructure to take performance measurements for PRs by taking a performance measurement of the project before the changes in the PR.

This will become really useful once we can switch this to a macOS runner and are then able to measure instructions executed by an executable (like swift-format supports using `swift-format --measure-instructions`) – instruction counting isn’t available on Linux as far as I could find out.

But even now, we can use this to track other metrics like binary size.