Do not assume windows_build_command is a single line
`windows_build_command` was prefixed with a `Invoke-Program` to ensure
we catch errors, but this then misses multiline commands. Move
`Invoke-Program` up into the default command and then assume that it is
passed if a non-default command is used.

Note that `$ErrorActionPreference = 'Stop'` is not an option here - it
only stops for cmdlets, not native commands.
`$PSNativeCommandUseErrorActionPreference = $true` allows stopping on
native commands too, but that was only introduced in PowerShell 7.3.
windowsservercore-ltsc2022 has 5.1 by default (though we could install a
newer if we wanted to).

Update PR testing in this repo to use `Invoke-Program` to then actually
catch failures.