Add a test package for CI
We currently have setup the workflows in this repository to test the various reusable workflows. The workflows that require a Swift package have created their own package as a pre-build command. While this worked so far it limits the kinds of packages that we can setup. This PR is committing a Swift Package to the `tests` folder which we can shape more complicated ways to test the various workflows better.