Rename install-swift-nightly.ps1 to install-swift-nightly-main.ps1 (#200)
This script is only used when Docker is enabled, and the workflow was using the wrong name, leading to build failures.

e.g. https://github.com/swiftlang/swift-build/actions/runs/19923980997/job/57119158460?pr=954

Easiest solution is to rename the file.

For Docker, both "nightly" and "nightly-main" tags exist, so we can just use "nightly-main" which is both more consistent with other platforms and easier to use across the board.