Fix error annotations
The previous method of excluding matrix entries results in complaints annotating the action run (even though the matrix produces the correct entries and _works_), but this noise gets in the way of real errors and may be bothersome.

Use a new technique (which amusingly is even simpler) to skip the whole matrix when the enabled flag is off for a group of builds.